requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic_settings
python-dotenv
passlib
python-jose[cryptography]
google-generativeai
loguru